+++ src/business/services/deposit.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 
 import { DepositEntity, DepositRepository } from '../../data/persistence';
-import { HistoryDto, PaginationDto } from '../../business/dtos';
+import { BalanceDto, DateRangeDto, PaginationDto } from '../../business/dtos';
 import { AccountService } from '.';
 
 @Injectable()
@@ -10,14 +10,23 @@ export class DepositService {
   constructor(private readonly depositRepository: DepositRepository, private readonly accountService: AccountService) {}
 
   createDeposit(deposit: DepositEntity): DepositEntity {
+    const newBalance = new BalanceDto();
+    newBalance.accountId = deposit.account.id;
+    newBalance.amount = deposit.amount;
+    this.accountService.addBalance(newBalance);
     return this.depositRepository.register(deposit);
   }
 
   deleteDeposit(depositId: string): boolean {
-    try{
-      this.depositRepository.delete(depositId);
-      return true;
-    } catch {
+    const current = this.depositRepository.findOneById(depositId);
+    if (current){
+      try{
+        this.depositRepository.delete(depositId);
+        return true;
+      } catch {
+        return false;
+      }
+    } else {
       return false;
     }
   }
@@ -25,10 +34,10 @@ export class DepositService {
   getHistory(
     accountId: string,
     pagination?: PaginationDto,
-    dataRange?: HistoryDto,
+    dataRange?: DateRangeDto,
   ): DepositEntity[] {
     if (dataRange){
-      return this.depositRepository.findByDataRange(accountId, dataRange.dateInit, dataRange.dateEnd, pagination);
+      return this.depositRepository.findByDataRange(accountId, dataRange?.dateInit, dataRange?.dateEnd, pagination);
     } else return this.depositRepository.findByAccountId(accountId, pagination);
   }
 }
