+++ src/business/services/transfer.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { TransferEntity, TransferRepository } from '../../data/persistence';
-import { HistoryDto, PaginationDto } from '../../business/dtos';
+import { DateRangeDto, PaginationDto } from '../../business/dtos';
 import { AccountService } from '.';
 
 @Injectable()
@@ -15,38 +15,43 @@ export class TransferService {
   getHistoryOut(
     accountId: string,
     pagination?: PaginationDto,
-    dataRange?: HistoryDto,
+    dataRange?: DateRangeDto,
   ): TransferEntity[] {
     if (dataRange){
-      return this.transferRepository.findOutcomeByDataRange(accountId, dataRange.dateInit, dataRange.dateEnd, pagination);
+      return this.transferRepository.findOutcomeByDataRange(accountId, dataRange?.dateInit, dataRange?.dateEnd, pagination);
     } else return this.transferRepository.findByOutcomeId(accountId, pagination);
   }
 
   getHistoryIn(
     accountId: string,
     pagination?: PaginationDto,
-    dataRange?: HistoryDto,
+    dataRange?: DateRangeDto,
   ): TransferEntity[] {
     if (dataRange){
-      return this.transferRepository.findIncomeByDataRange(accountId, dataRange.dateInit, dataRange.dateEnd, pagination);
+      return this.transferRepository.findIncomeByDataRange(accountId, dataRange?.dateInit, dataRange?.dateEnd, pagination);
     } else return this.transferRepository.findByIncomeId(accountId, pagination);
   }
 
   getHistory(
     accountId: string,
     pagination?: PaginationDto,
-    dataRange?: HistoryDto,
+    dataRange?: DateRangeDto,
   ): TransferEntity[] {
     if (dataRange){
-      return this.transferRepository.findIncomeByDataRange(accountId, dataRange.dateInit, dataRange.dateEnd, pagination).concat(this.transferRepository.findOutcomeByDataRange(accountId, dataRange.dateInit, dataRange.dateEnd, pagination));
+      return this.transferRepository.findIncomeByDataRange(accountId, dataRange?.dateInit, dataRange?.dateEnd, pagination).concat(this.transferRepository.findOutcomeByDataRange(accountId, dataRange?.dateInit, dataRange?.dateEnd, pagination));
     } else return this.transferRepository.findByIncomeId(accountId, pagination).concat(this.transferRepository.findByOutcomeId(accountId, pagination));
   }
 
   deleteTransfer(transferId: string): boolean {
-    try{
-      this.transferRepository.delete(transferId);
-      return true;
-    } catch {
+    const current = this.transferRepository.findOneById(transferId);
+    if (current){
+      try{
+        this.transferRepository.delete(transferId);
+        return true;
+      } catch {
+        return false;
+      }
+    } else {
       return false;
     }
   }
